requirements.txt
Flask==2.2.2
pytesseract==0.3.10
Pillow==9.4.0
Flask-Cors==3.0.10
requests==2.28.2